At TargetSelectActivity with new DB adaptations

+++ app/src/main/java/edu/wsu/lar/airpact_fire/data/realm/model/Coordinate.java
@@ -4,5 +4,5 @@ import io.realm.RealmObject;
 
 public class Coordinate extends RealmObject {
 
-    public float x, y;
+    public double x, y;
 }
